Fix more memory leaks
I think this is my last commit trying to use OwnedPtr where needed. It's
just too confusing. I fear no one will be able to understand how to work
on the parser, and our reliance on AddressSanitizer to catch bugs is
worrisome -- what if it misses one or we don't fuzz enough?

In the next commit, I'd like to wrap all nodes with SharedPtr and then to
benchmark the difference in speed. I'm hoping the overhead is marginal.