Fix bugs parsing word array with interpolation
Honestly, this state machine could use a lot of work. It's very
spaghetti-code and hard to work on because the states & transitions are
intermingled with the state of the buffer.

Potential improvement: explicitly name the buffer states as well, so,
in addition to `InProgress`, we would also get `InProgressBufferEmpty`
or somesuch.

Doing it that way would allow us to avoid checking the buffer size and
would make the transitions a bit clearer, I think.